+++ packages/orq-rc/RELEASES.md
@@ -1339,3 +1339,13 @@ Based on:
 - [typescript v3.4.0-rc.21] packages/orq-rc
 ### Releases
 - [NPM v3.4.0-rc.21] https://www.npmjs.com/package/@orq-ai/node/v/3.4.0-rc.21 - packages/orq-rc
+
+## 2025-04-15 11:50:34
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.531.5 (2.570.4) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v3.4.0-rc.22] packages/orq-rc
+### Releases
+- [NPM v3.4.0-rc.22] https://www.npmjs.com/package/@orq-ai/node/v/3.4.0-rc.22 - packages/orq-rc